requirements.txt
groq==0.29.0
huggingface-hub==0.33.2
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-groq==0.3.5
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydeck==0.9.1
pypdf==5.7.0
python-dotenv==1.1.1
pyyaml==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
streamlit==1.46.1
tokenizers==0.21.2
toml==0.10.2
transformers==4.53.1